backend/requirements.txt
# Google GenAI SDK & ADK
google-genai>=1.0.0
google-adk>=0.3.0

# Web framework
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
websockets>=12.0

# Utilities
python-dotenv>=1.0.0
